+++ datalad/downloaders/providers.py
@@ -20,8 +20,12 @@ from collections import OrderedDict
 
 from .base import NoneAuthenticator, NotImplementedAuthenticator
 
-from .http import HTMLFormAuthenticator, HTTPBasicAuthAuthenticator, HTTPDigestAuthAuthenticator
-from .http import HTTPDownloader
+from .http import (
+    HTMLFormAuthenticator, HTTPBasicAuthAuthenticator,
+    HTTPDigestAuthAuthenticator,
+    HTTPBearerTokenAuthenticator,
+    HTTPDownloader,
+)
 from .s3 import S3Authenticator, S3Downloader
 from ..support.configparserinc import SafeConfigParserWithIncludes
 from ..support.external_versions import external_versions
@@ -40,8 +44,10 @@ AUTHENTICATION_TYPES = {
     'http_auth': HTTPBasicAuthAuthenticator,
     'http_basic_auth': HTTPBasicAuthAuthenticator,
     'http_digest_auth': HTTPDigestAuthAuthenticator,
+    'bearer_token': HTTPBearerTokenAuthenticator,
     'aws-s3': S3Authenticator,  # TODO: check if having '-' is kosher
     'nda-s3': S3Authenticator,
+    'loris-token': HTTPBearerTokenAuthenticator,
     'xnat': NotImplementedAuthenticator,
     'none': NoneAuthenticator,
 }