+++ CHANGELOG.md
@@ -9,6 +9,17 @@ This is a high level and scarce summary of the changes between releases.
 We would recommend to consult log of the 
 [DataLad git repository](http://github.com/datalad/datalad) for more details.
 
+## 0.10.1 (Jun 17, 2018) -- OHBM polish
+
+The is a minor bugfix release.
+
+### Fixes
+
+- Be able to use backports.lzma as a drop-in replacement for pyliblzma.
+- Give help when not specifying a procedure name in `run-procedure`.
+- Abort early when a downloader received no filename.
+- Avoid `rerun` error when trying to unlock non-available files.
+
 ## 0.10.0 (Jun 09, 2018) -- The Release
 
 This release is a major leap forward in metadata support.